[FIX] udes_priorities: Propagate priority sequence on write instead of compute.
On large databases with many pickings, Odoo spends a very long time preparing
the pickings for compute and this can lead to hour(s) long wait times
if moving a priority sequence up or down. This change just propagates to the relevant
pickings on write, which means the precomputation is no longer necessary
and the speed improvement is orders of magnitude faster.

The only downside, is that if the picking priority changes, then we need to handle
updating the sequence there with a write/create hook.

I have removed the `u_priority_id` computed field along with this as it
is not needed.

SE-1268

Signed-off-by: Peter Alabaster <peter.alabaster@unipart.com>