[TEST] udes_security: Add doctype to html test file
SE-1554

Signed-off-by: Kevin Dwyer <kevin.dwyer@unipart.com>

So that it is recognised as html. Without this change
`TestAllowedFileType.test_cannot_change_attachment_file_type_to_invalid_one`
fails locally for me, although it does not fail in CI.

We don't use this file to test that html files are correctly identified,
so adding the doctype should affect the quality of our tests.