[ADD] udes_stock_routing: Prevent push on drop rule when move has linked future moves.
This is because a 2 stage outbound setup likely coincides with a 2 stage
inbound setup, where the intermediate locations will be the same.

We will have a configured push on drop rule for 2 stage inbound and
we want to prevent outbound moves to the intermediate location triggering
the push on drop rule.

The approach we are taking is to check if the move has any followup moves,
and if so - it will skip push on drop.

There _could_ be issues with this approach but it seems fairly robust for now.
SE-1701

Signed-off-by: Peter Alabaster <peter.alabaster@unipart.com>