[IMP] udes_stock_routing: Unreserving split picks
SE-2144

Signed-off-by: Kevin Dwyer <kevin.dwyer@unipart.com>

When split picks are unreserved, we want to return the system to its
state prior to the splitting, insofar as this is possible.

Moves from unreserved split picks should be put back into the original
pick. If the original pick cannot be amended then a new pick of that
picking type should be created from the moves.

Generated bulk picks left without moves after unreservation should be
marked for deletion.

This prevents the system from simply re-assigning the generated split
picks on re-reservation, as the related moves may have been wrong or
undesirable.