DRAFT: Hotfix 27 10 merge (#968)
* [FIX] udes_stock_cron: Do not reserve stock in draft batches

SE-2137

Signed-off-by: Kevin Dwyer <kevin.dwyer@unipart.com>

The add_processed_pickings method adds all the pickings in a batch to
the processed collection, and this messes up the comparison in
reserve_stock_for_picking_type.

Don't process if the current picking is in the processed collection.

* [FIX] udes_stock_picking_batch: Batch compute priority

Make sure all the priorities are str, in some cases pickings are
returning False, meanwhile ignore them.

Signed-off-by: Miquel PS <miquel.palahi.sitges@unipart.io>

* [FIX] udes_stock_picking_batch: Batch compute priority default value

Set default value as str.

Signed-off-by: Miquel PS <miquel.palahi.sitges@unipart.io>

* [ADD] udes_stock_routing: Copying sequence when splitting the pick (#965)

Signed-off-by: Armand Cela <armand.cela@unipart.io>

* [IMP] udes_stock: Add filter hooks for Quant._gather

SE-2144

Signed-off-by: Kevin Dwyer <kevin.dwyer@unipart.com>

Enable arbitrary filtering of gathered quants.

* [IMP] udes_stock_routing: Unreserving split picks

SE-2144

Signed-off-by: Kevin Dwyer <kevin.dwyer@unipart.com>

When split picks are unreserved, we want to return the system to its
state prior to the splitting, insofar as this is possible.

Moves from unreserved split picks should be put back into the original
pick. If the original pick cannot be amended then a new pick of that
picking type should be created from the moves.

Generated bulk picks left without moves after unreservation should be
marked for deletion.

This prevents the system from simply re-assigning the generated split
picks on re-reservation, as the related moves may have been wrong or
undesirable.

* [FIX] Always return res in _do_unreserve hook

* [ADD] udes_stock_routing: Skipping some unit tests till DEBT is reviewed

Skipping split unit tests, there is a story which currently is being worked and these tests needs reviewing
after the story is done

Signed-off-by: Armand Cela <armand.cela@unipart.io>

---------

Signed-off-by: Miquel PS <miquel.palahi.sitges@unipart.io>
Signed-off-by: Armand Cela <armand.cela@unipart.io>
Co-authored-by: Kevin Dwyer <kevin.dwyer@unipart.com>
Co-authored-by: Miquel PS <miquel.palahi.sitges@unipart.io>
Co-authored-by: Peter Alabaster <peter.alabaster@unipart.com>